feat(context-pad): make usage of auto-place configurable
Library users may opt-out of auto place from
context pad by specifying the configuration option
contextPad.autoPlace = false.

Related to #563
