+++ lib/features/context-pad/ContextPadProvider.js
@@ -14,9 +14,11 @@ var assign = require('lodash/object/assign'),
 /**
  * A provider for BPMN 2.0 elements context pad
  */
-function ContextPadProvider(injector, eventBus, contextPad, modeling,
+function ContextPadProvider(config, injector, eventBus, contextPad, modeling,
                             elementFactory, connect, create, popupMenu,
                             canvas, rules, translate) {
+
+  config = config || {};
 
   contextPad.registerProvider(this);
 
@@ -31,7 +33,10 @@ function ContextPadProvider(injector, eventBus, contextPad, modeling,
   this._canvas  = canvas;
   this._rules = rules;
   this._translate = translate;
-  this._autoPlace = injector.get('autoPlace', false);
+
+  if (config.autoPlace !== false) {
+    this._autoPlace = injector.get('autoPlace', false);
+  }
 
   eventBus.on('create.end', 250, function(event) {
     var shape = event.context.shape;
@@ -49,6 +54,7 @@ function ContextPadProvider(injector, eventBus, contextPad, modeling,
 }
 
 ContextPadProvider.$inject = [
+  'config.contextPad',
   'injector',
   'eventBus',
   'contextPad',
